Update .travis.yml
Hi @ajhynes7 I was trying to get the tests up and running and, I'm not really sure, but I guess this `$stage` parameter is unneeded here, as `run_stages.sh` don't read this. It's probably working because it's an empty string but maybe it's more clear with the fix, specially because it's the _how-to-test_ documentation.

Best regards!